+++ packages/nuxt-auth-module/types.d.ts
@@ -1,0 +1,7 @@
+import type { AuthModule } from './lib/auth-module';
+
+declare module '@nuxt/types' {
+	interface Context {
+		$auth: AuthModule;
+	}
+}