+++ packages/vue-form-composition/src/control/index.ts
@@ -1,2 +1,0 @@
-export * from './form-control-props';
-export * from './form-control';