+++ packages/vue-form-composition/src/index.ts
@@ -1,4 +1,6 @@
-export * from './control';
-export * from './rule';
+export * from './state';
+export * from './validator';
 export * from './props';
 export * from './use-form';
+export * from './use-form-model';
+export type { FormType, FormDefinition } from './types';