+++ packages/vue-notification-composition/src/constants.ts
@@ -1,0 +1,1 @@
+export const TIMEOUT_DELAY = 5000;