+++ packages/vuetify-util/src/select/index.ts
@@ -1,0 +1,2 @@
+export * from './select-props';
+export * from './select';