+++ CHANGELOG.md
@@ -3,6 +3,275 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/rhangai/node-web-packages/compare/v0.2.4...v0.2.5) (2021-06-10)
+
+
+### Bug Fixes
+
+* Lifecycle scripts ([ac63af0](https://github.com/rhangai/node-web-packages/commit/ac63af0ee481dd8728b7bc30d7c1c54e426c69f5))
+
+
+
+
+
+## [0.2.4](https://github.com/rhangai/node-web-packages/compare/v0.2.3...v0.2.4) (2021-06-07)
+
+
+### Bug Fixes
+
+* Typings for useForm ([f60d782](https://github.com/rhangai/node-web-packages/commit/f60d7829c7b4b549572f2d13ae484da5aa753470))
+
+
+
+
+
+## [0.2.3](https://github.com/rhangai/node-web-packages/compare/v0.2.2...v0.2.3) (2021-06-05)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.2.2](https://github.com/rhangai/node-web-packages/compare/v0.2.1...v0.2.2) (2021-06-05)
+
+
+### Bug Fixes
+
+* Typing of optional definition ([6f0e90c](https://github.com/rhangai/node-web-packages/commit/6f0e90cc7c62cd9a8f8c7d9d122f613e96e45776))
+
+
+
+
+
+## [0.2.1](https://github.com/rhangai/node-web-packages/compare/v0.2.0...v0.2.1) (2021-06-05)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+# [0.2.0](https://github.com/rhangai/node-web-packages/compare/v0.1.24...v0.2.0) (2021-06-05)
+
+
+### Features
+
+* Removed submit configuration ([a1d66e8](https://github.com/rhangai/node-web-packages/commit/a1d66e809ae21802e7cd97c8d92f69f69ed5a18c))
+
+
+
+
+
+## [0.1.24](https://github.com/rhangai/node-web-packages/compare/v0.1.23...v0.1.24) (2021-06-05)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.23](https://github.com/rhangai/node-web-packages/compare/v0.1.22...v0.1.23) (2021-06-05)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.22](https://github.com/rhangai/node-web-packages/compare/v0.1.21...v0.1.22) (2021-06-05)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.21](https://github.com/rhangai/node-web-packages/compare/v0.1.20...v0.1.21) (2021-06-05)
+
+
+### Bug Fixes
+
+* Typings of onValue ([de5d41d](https://github.com/rhangai/node-web-packages/commit/de5d41dd992e0728018843f53117673f165d1ea7))
+
+
+
+
+
+## [0.1.20](https://github.com/rhangai/node-web-packages/compare/v0.1.19...v0.1.20) (2021-06-05)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.19](https://github.com/rhangai/node-web-packages/compare/v0.1.18...v0.1.19) (2021-06-03)
+
+
+### Bug Fixes
+
+* Export of FormDefinition ([5a75130](https://github.com/rhangai/node-web-packages/commit/5a7513065d1251cd501aca192f42122c36953b0e))
+
+
+
+
+
+## [0.1.18](https://github.com/rhangai/node-web-packages/compare/v0.1.17...v0.1.18) (2021-06-03)
+
+
+### Bug Fixes
+
+* FormDefinition type to remove readonly ([4c641fd](https://github.com/rhangai/node-web-packages/commit/4c641fd9151c44a6fa06c6ce639575f9cf309c26))
+
+
+
+
+
+## [0.1.17](https://github.com/rhangai/node-web-packages/compare/v0.1.16...v0.1.17) (2021-06-03)
+
+
+### Bug Fixes
+
+* Await on options.validate ([bed8a84](https://github.com/rhangai/node-web-packages/commit/bed8a8430fffef45b28f172144fedd41ca5aa307))
+
+
+
+
+
+## [0.1.16](https://github.com/rhangai/node-web-packages/compare/v0.1.15...v0.1.16) (2021-06-03)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.15](https://github.com/rhangai/node-web-packages/compare/v0.1.14...v0.1.15) (2021-06-03)
+
+
+### Bug Fixes
+
+* Delay when resolving confirmations/notifications ([86660f0](https://github.com/rhangai/node-web-packages/commit/86660f0e00cfc4faa5fab238b7ccbbdef3a8a562))
+
+
+
+
+
+## [0.1.14](https://github.com/rhangai/node-web-packages/compare/v0.1.13...v0.1.14) (2021-06-03)
+
+
+### Bug Fixes
+
+* Name of user variable ([528df55](https://github.com/rhangai/node-web-packages/commit/528df55c887889a5e726af6d7a9fdacaaceeb777))
+
+
+
+
+
+## [0.1.13](https://github.com/rhangai/node-web-packages/compare/v0.1.12...v0.1.13) (2021-06-02)
+
+
+### Bug Fixes
+
+* UseSubmitOptions type ([06c75bf](https://github.com/rhangai/node-web-packages/commit/06c75bf03d8c8bd90c8848b08514b7040fd5839b))
+
+
+
+
+
+## [0.1.12](https://github.com/rhangai/node-web-packages/compare/v0.1.11...v0.1.12) (2021-06-02)
+
+
+### Bug Fixes
+
+* Form now allow for null on nested objects ([cb37030](https://github.com/rhangai/node-web-packages/commit/cb37030cd38bbeb5e3e1842f98fd73190a70a04e))
+
+
+
+
+
+## [0.1.11](https://github.com/rhangai/node-web-packages/compare/v0.1.10...v0.1.11) (2021-06-02)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.10](https://github.com/rhangai/node-web-packages/compare/v0.1.9...v0.1.10) (2021-06-01)
+
+
+### Bug Fixes
+
+* Export of $auth on nuxt context ([9f43b71](https://github.com/rhangai/node-web-packages/commit/9f43b7187bb1b17194c67ed5f162b65d4818b6ae))
+
+
+
+
+
+## [0.1.9](https://github.com/rhangai/node-web-packages/compare/v0.1.8...v0.1.9) (2021-06-01)
+
+
+### Bug Fixes
+
+* Exports of nuxt-auth-module ([1d22ee5](https://github.com/rhangai/node-web-packages/commit/1d22ee504c47be80dec6176a550f2f8c41beea05))
+
+
+
+
+
+## [0.1.8](https://github.com/rhangai/node-web-packages/compare/v0.1.7...v0.1.8) (2021-06-01)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.7](https://github.com/rhangai/node-web-packages/compare/v0.1.6...v0.1.7) (2021-06-01)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.6](https://github.com/rhangai/node-web-packages/compare/v0.1.5...v0.1.6) (2021-06-01)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.5](https://github.com/rhangai/node-web-packages/compare/v0.1.4...v0.1.5) (2021-06-01)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.4](https://github.com/rhangai/node-web-packages/compare/v0.1.3...v0.1.4) (2021-05-31)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.3](https://github.com/rhangai/node-web-packages/compare/v0.1.2...v0.1.3) (2021-05-31)
+
+
+### Bug Fixes
+
+* Request result type ([34d407d](https://github.com/rhangai/node-web-packages/commit/34d407dad68ff66f6b4e6a6b2eaadd5f5bdaa453))
+
+
+
+
+
 ## [0.1.2](https://github.com/rhangai/node-web-packages/compare/v0.1.1...v0.1.2) (2021-05-28)
 
 **Note:** Version bump only for package root