+++ packages/common/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/rhangai/node-web-packages/compare/v0.2.4...v0.2.5) (2021-06-10)
+
+
+### Bug Fixes
+
+* Lifecycle scripts ([ac63af0](https://github.com/rhangai/node-web-packages/commit/ac63af0ee481dd8728b7bc30d7c1c54e426c69f5))
+
+
+
+
+
+# [0.2.0](https://github.com/rhangai/node-web-packages/compare/v0.1.24...v0.2.0) (2021-06-05)
+
+**Note:** Version bump only for package @rhangai/web-common
+
+
+
+
+
+## [0.1.12](https://github.com/rhangai/node-web-packages/compare/v0.1.11...v0.1.12) (2021-06-02)
+
+**Note:** Version bump only for package @rhangai/web-common
+
+
+
+
+
+## [0.1.11](https://github.com/rhangai/node-web-packages/compare/v0.1.10...v0.1.11) (2021-06-02)
+
+**Note:** Version bump only for package @rhangai/web-common
+
+
+
+
+
 # [0.1.0](https://github.com/rhangai/node-web-packages/compare/v0.0.8...v0.1.0) (2021-05-27)
 
 **Note:** Version bump only for package @rhangai/web-common