+++ packages/vue-form-composition/CHANGELOG.md
@@ -3,6 +3,155 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/rhangai/node-web-packages/compare/v0.2.4...v0.2.5) (2021-06-10)
+
+
+### Bug Fixes
+
+* Lifecycle scripts ([ac63af0](https://github.com/rhangai/node-web-packages/commit/ac63af0ee481dd8728b7bc30d7c1c54e426c69f5))
+
+
+
+
+
+## [0.2.4](https://github.com/rhangai/node-web-packages/compare/v0.2.3...v0.2.4) (2021-06-07)
+
+
+### Bug Fixes
+
+* Typings for useForm ([f60d782](https://github.com/rhangai/node-web-packages/commit/f60d7829c7b4b549572f2d13ae484da5aa753470))
+
+
+
+
+
+## [0.2.3](https://github.com/rhangai/node-web-packages/compare/v0.2.2...v0.2.3) (2021-06-05)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
+## [0.2.2](https://github.com/rhangai/node-web-packages/compare/v0.2.1...v0.2.2) (2021-06-05)
+
+
+### Bug Fixes
+
+* Typing of optional definition ([6f0e90c](https://github.com/rhangai/node-web-packages/commit/6f0e90cc7c62cd9a8f8c7d9d122f613e96e45776))
+
+
+
+
+
+## [0.2.1](https://github.com/rhangai/node-web-packages/compare/v0.2.0...v0.2.1) (2021-06-05)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
+# [0.2.0](https://github.com/rhangai/node-web-packages/compare/v0.1.24...v0.2.0) (2021-06-05)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
+## [0.1.23](https://github.com/rhangai/node-web-packages/compare/v0.1.22...v0.1.23) (2021-06-05)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
+## [0.1.21](https://github.com/rhangai/node-web-packages/compare/v0.1.20...v0.1.21) (2021-06-05)
+
+
+### Bug Fixes
+
+* Typings of onValue ([de5d41d](https://github.com/rhangai/node-web-packages/commit/de5d41dd992e0728018843f53117673f165d1ea7))
+
+
+
+
+
+## [0.1.20](https://github.com/rhangai/node-web-packages/compare/v0.1.19...v0.1.20) (2021-06-05)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
+## [0.1.19](https://github.com/rhangai/node-web-packages/compare/v0.1.18...v0.1.19) (2021-06-03)
+
+
+### Bug Fixes
+
+* Export of FormDefinition ([5a75130](https://github.com/rhangai/node-web-packages/commit/5a7513065d1251cd501aca192f42122c36953b0e))
+
+
+
+
+
+## [0.1.18](https://github.com/rhangai/node-web-packages/compare/v0.1.17...v0.1.18) (2021-06-03)
+
+
+### Bug Fixes
+
+* FormDefinition type to remove readonly ([4c641fd](https://github.com/rhangai/node-web-packages/commit/4c641fd9151c44a6fa06c6ce639575f9cf309c26))
+
+
+
+
+
+## [0.1.12](https://github.com/rhangai/node-web-packages/compare/v0.1.11...v0.1.12) (2021-06-02)
+
+
+### Bug Fixes
+
+* Form now allow for null on nested objects ([cb37030](https://github.com/rhangai/node-web-packages/commit/cb37030cd38bbeb5e3e1842f98fd73190a70a04e))
+
+
+
+
+
+## [0.1.11](https://github.com/rhangai/node-web-packages/compare/v0.1.10...v0.1.11) (2021-06-02)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
+## [0.1.5](https://github.com/rhangai/node-web-packages/compare/v0.1.4...v0.1.5) (2021-06-01)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
+## [0.1.4](https://github.com/rhangai/node-web-packages/compare/v0.1.3...v0.1.4) (2021-05-31)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
+## [0.1.3](https://github.com/rhangai/node-web-packages/compare/v0.1.2...v0.1.3) (2021-05-31)
+
+**Note:** Version bump only for package @rhangai/vue-form-composition
+
+
+
+
+
 ## [0.1.2](https://github.com/rhangai/node-web-packages/compare/v0.1.1...v0.1.2) (2021-05-28)
 
 **Note:** Version bump only for package @rhangai/vue-form-composition